Moved A LOT of stuff around...created a filters folder to house all funcs that return an image
