+++ porespy/__init__.py
@@ -10,3 +10,4 @@ from . import simulations
 from . import metrics
 from . import generators
 from . import binarization
+from . import filters
