+++ porespy/filters/__init__.py
@@ -1,0 +1,11 @@
+r"""
+===============================================================================
+Filters
+===============================================================================
+
+
+"""
+from .__funcs__ import apply_chords, apply_chords_3D
+from .__funcs__ import local_thickness, flood
+from .__funcs__ import find_disconnected_voxels, trim_extrema
+from .__funcs__ import trim_floating_solid, fill_blind_pores
