+++ porespy/metrics/__init__.py
@@ -7,8 +7,6 @@ A suite of tools for determine key metrics about an image, including:
 ``porosity`` - Quickly determines the ratio of void voxels to total voxels
 of the image.
 """
-from .__funcs__ import feature_size_distribution, porosity
-from .__funcs__ import representative_elementary_volume
-from .__funcs__ import apply_chords, apply_chords_3D, chord_length_distribution
+from .__funcs__ import representative_elementary_volume, porosity
+from .__funcs__ import chord_length_distribution
 from .__funcs__ import two_point_correlation_bf, pore_size_density
-from .__funcs__ import local_thickness
