+++ porespy/tools/__funcs__.py
@@ -250,57 +250,6 @@ def randomize_colors(im, keep_vals=[0]):
     return im_new
 
 
-@jit
-def flood(im, mode='max'):
-    r"""
-    Floods/fills each region in an image with a single value based on the
-    specific values in that region.  The ```mode``` argument is used to
-    determine how the value is calculated.  A region is defined as a connected
-    cluster of voxels surrounded by 0's for False's.
-
-    Parameters
-    ----------
-    im : array_like
-        An ND image with isolated regions containing 0's elsewhere.
-
-    mode : string
-        Specifies how to determine which value should be used to flood each
-        region.  Options are:
-
-    *'max'* : Floods each region with the local maximum in that region
-
-    *'min'* : Floods each region the local minimum in that region
-
-    *'size'* : Floods each region with the size of that region
-
-    Returns
-    -------
-    An ND-array the same size as ```im``` with new values placed in each
-    forground voxel based on the ```mode```.
-
-    """
-    labels, N = spim.label(im)
-    I = im.flatten()
-    L = labels.flatten()
-    if mode.startswith('max'):
-        V = sp.zeros(shape=N+1, dtype=float)
-        for i in range(len(L)):
-            if V[L[i]] < I[i]:
-                V[L[i]] = I[i]
-    elif mode.startswith('min'):
-        V = sp.ones(shape=N+1, dtype=float)*sp.inf
-        for i in range(len(L)):
-            if V[L[i]] > I[i]:
-                V[L[i]] = I[i]
-    elif mode.startswith('size'):
-        V = sp.zeros(shape=N+1, dtype=int)
-        for i in range(len(L)):
-            V[L[i]] += 1
-    im_flooded = sp.reshape(V[labels], newshape=im.shape)
-    im_flooded = im_flooded*im
-    return im_flooded
-
-
 def make_contiguous(im):
     r"""
     Take an image with arbitrary greyscale values and adjust them to ensure
@@ -410,79 +359,3 @@ def get_border(shape, thickness=1, mode='edges'):
             border[0::, t:-t, 0::] = False
             border[0::, 0::, t:-t] = False
     return border
-
-
-def find_disconnected_voxels(im, conn=None):
-    r"""
-    This identifies all pore (or solid) voxels that are not connected to the
-    edge of the image.  This can be used to find blind pores, or remove
-    artifacts such as solid phase voxels that are floating in space.
-
-    Parameters
-    ----------
-    im : ND-image
-        A Boolean image, with True values indicating the phase for which
-        disconnected voxels are sought.
-
-    conn : int
-        For 2D the options are 4 and 8 for square and diagonal neighbors, while
-        for the 3D the options are 6 and 26, similarily for square and diagonal
-        neighbors.  The default is max
-
-    Returns
-    -------
-    An ND-image the same size as ``im``, with True values indicating voxels of
-    the phase of interest (i.e. True values in the original image) that are
-    not connected to the outer edges.
-
-    Notes
-    -----
-    The returned array (e.g. ``holes``) be used to trim blind pores from
-    ``im`` using: ``im[holes] = False``
-
-    """
-    if im.ndim == 2:
-        if conn == 4:
-            strel = disk(1)
-        elif conn in [None, 8]:
-            strel = square(3)
-    elif im.ndim == 3:
-        if conn == 6:
-            strel = ball(1)
-        elif conn in [None, 26]:
-            strel = cube(3)
-    labels, N = spim.label(input=im, structure=strel)
-    holes = clear_border(labels=labels) > 0
-    return holes
-
-
-def trim_extrema(im, h, mode='maxima'):
-    r"""
-    This trims local extrema in greyscale values by a specified amount,
-    essentially decapitating peaks or flooding valleys, or both.
-
-    Parameters
-    ----------
-    im : ND-array
-        The image whose extrema are to be removed
-
-    h : scalar
-        The height to remove from each peak or fill in each valley
-
-    mode : string {'maxima' | 'minima' | 'extrema'}
-        Specifies whether to remove maxima or minima or both
-
-    Returns
-    -------
-    A copy of the input image with all the peaks and/or valleys removed.
-
-    Notes
-    -----
-    This function is referred to as **imhmax** or **imhmin** in Mablab.
-    """
-    result = im
-    if mode in ['maxima', 'extrema']:
-        result = reconstruction(seed=im - h, mask=im, method='dilation')
-    elif mode in ['minima', 'extrema']:
-        result = reconstruction(seed=im + h, mask=im, method='erosion')
-    return result
