+++ porespy/tools/__init__.py
@@ -1,4 +1,4 @@
 from .__funcs__ import randomize_colors, make_contiguous, get_border
-from .__funcs__ import flood, binary_opening_fast, get_slice
+from .__funcs__ import binary_opening_fast, get_slice
 from .__funcs__ import extend_slice, extract_subsection
-from .__funcs__ import find_disconnected_voxels, find_outer_region
+from .__funcs__ import find_outer_region
